requirements.txt
# Core PyTorch dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Deep learning and ML
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data handling
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
requests>=2.31.0
urllib3>=2.0.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# GPU monitoring (optional)
nvidia-ml-py3>=7.352.0
psutil>=5.9.0
